move TagDescriptionEditor into modal
